requirements.txt
Django
djangorestframework
psycopg2
pygments
httpie
coreapi
coreapi-cli
